Move id generation into internal/
