+++ packages/localize/package.json
@@ -26,18 +26,10 @@
     "!/src/tests/",
     "/src_client/",
     "!/src_client/tests/",
-    "/lib/",
-    "!/lib/tests/",
-    "!/lib/.tsbuildinfo",
     "/lit-localize.d.ts",
     "/lit-localize.js",
     "/lit-localize.js.map",
-    "/id-generation.d.ts",
-    "/id-generation.js",
-    "/id-generation.js.map",
-    "/fnv1a64.d.ts",
-    "/fnv1a64.js",
-    "/fnv1a64.js.map",
+    "/internal/",
     "/config.schema.json"
   ],
   "dependencies": {
